+++ packages/core/docs/changelog.md
@@ -2,7 +2,7 @@
 
 ## v0.0.3
 
-- Added: FactoryInstance class
+- Added: Factory class
 - Added: Base64Url class
 - Added: Converter.bytesToBase64Url class
 - Added: Converter.base64UrlToBytes class